requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
Flask_Bcrypt==1.0.1
Flask_HTMLmin==2.2.1
Flask_Login==0.6.3
python-dotenv==1.0.1
Requests==2.31.0
stripe==9.5.0